requirements-managed.txt
openai>=1.79.0
streamlit==1.56.0
fastapi==0.135.3
uvicorn==0.44.0
pydantic==2.12.5
PyYAML==6.0.3
requests==2.33.1
pypdf==6.9.2
numpy==2.2.6
faiss-cpu==1.13.0
sentence-transformers==5.1.2
transformers==4.57.1
